This is a reimplementation of the optimizations initially proposed in PR https://github.com/ANTsX/ANTs/pull/1452

With the help of the ITK discourse forum in post https://discourse.itk.org/t/how-to-use-itk-multi-threading-primitives-to-do-parallel-execution-of-loop-iteration/5531/ I have managed to rework my previous proof of concept implementation into a generic solution that uses the ITK multi-threading primitive ParallelizeImageRegion.

During my limited testing (using only two image inputs) I have seen the execution time being reduced to around 60-70% of the original implementations execution time. The execution times for the two tested image inputs: image 1 was reduced from 96 seconds to 60 seconds (see below) and image 2 from 272 seconds to 168 seconds.

I have validated the implementation using the following:

Tests where done on the cpu:
```
lscpu | grep "Model name"
Model name:                      Intel(R) Xeon(R) CPU E5-2686 v4 @ 2.30GHz
```

Preparations:
```
wget https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0303:anat:sub-0303_T1w.nii.gz

mkdir master
ANTSPATH=./bin_old/ time ./bin_old/Scripts/antsAtroposN4.sh -d 3 -u 0 -c 3 -m 3 -n 3 -a sub-0303:anat:sub-0303_T1w.nii.gz -x sub-0303:anat:sub-0303_T1w.nii.gz -o master/
```

using a compiled version of ANTs before the code changes:
```
mkdir old_sub
time ./bin_old/Atropos -d 3 -x sub-0303:anat:sub-0303_T1w.nii.gz -c [ 3,0.0 ] -a master/Segmentation0N4.nii.gz --verbose 1 -i PriorProbabilityImages[ 3,master/SegmentationPosteriors%d.nii.gz,0.0 ] -k Gaussian -m [ 0.1,1x1x1 ] -o [ old_sub/Segmentation.nii.gz,old_sub/SegmentationPosteriors%d.nii.gz ] -r 0 -p Socrates[ 1 ]

Elapsed time: 96.0797

md5sum old_sub/*
fc6148a95f77d073768cbd46a18e340f  old_sub/Segmentation.nii.gz
38517eecae8745acbffe034f7ba71e10  old_sub/SegmentationPosteriors1.nii.gz
8974621586bf2844f4ef372dfcae96c2  old_sub/SegmentationPosteriors2.nii.gz
6c79422e44b49225490eb42f615cd1e0  old_sub/SegmentationPosteriors3.nii.gz
```

using a compiled version of ANTs after the code changes:
```
mkdir new_opt_sub
time ./bin_new/Atropos -d 3 -x sub-0303:anat:sub-0303_T1w.nii.gz -c [ 3,0.0 ] -a master/Segmentation0N4.nii.gz --verbose 1 -i PriorProbabilityImages[ 3,master/SegmentationPosteriors%d.nii.gz,0.0 ] -k Gaussian -m [ 0.1,1x1x1 ] -o [ new_opt_sub/Segmentation.nii.gz,new_opt_sub/SegmentationPosteriors%d.nii.gz ] -r 0 -p Socrates[ 1 ]

Elapsed time: 59.7719

md5sum new_opt_sub/*
fc6148a95f77d073768cbd46a18e340f  new_opt_sub/Segmentation.nii.gz
38517eecae8745acbffe034f7ba71e10  new_opt_sub/SegmentationPosteriors1.nii.gz
8974621586bf2844f4ef372dfcae96c2  new_opt_sub/SegmentationPosteriors2.nii.gz
6c79422e44b49225490eb42f615cd1e0  new_opt_sub/SegmentationPosteriors3.nii.gz
```

The md5sums confirms that the new implementation produces the exact same output as the before the changes.