@ntustison 2022-12-08
This is great.  However, I would prefer to see a few more subjects processed where the output doesn't change before merging.