@br-cpvc 2022-12-08
@ntustison Sure, good idea. I have tested 5 more images with the following:

```
wget https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0010:anat:sub-0010_T1w.nii.gz https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0086:anat:sub-0086_T1w.nii.gz https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0101:anat:sub-0101_T1w.nii.gz https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0109:anat:sub-0109_T1w.nii.gz https://openneuro.org/crn/datasets/ds004288/snapshots/1.0.1/files/sub-0114:anat:sub-0114_T1w.nii.gz

for sub in sub-0010 sub-0086 sub-0101 sub-0109 sub-0114; do \
export ver=old && mkdir -p $sub/$ver && echo "$sub $ver" && ./bin_$ver/Atropos -d 3 -x "${sub}:anat:${sub}_T1w.nii.gz" -c [ 3,0.0 ] -a "${sub}:anat:${sub}_T1w.nii.gz" -k Gaussian -m [ 0.1,1x1x1 ] -o $sub/$ver/Segmentation.nii.gz -r 0 -p Socrates[ 1 ] -i otsu[3] --verbose 1 | grep "Elapsed time:" && \
export ver=new && mkdir -p $sub/$ver && echo "$sub $ver" && ./bin_$ver/Atropos -d 3 -x "${sub}:anat:${sub}_T1w.nii.gz" -c [ 3,0.0 ] -a "${sub}:anat:${sub}_T1w.nii.gz" -k Gaussian -m [ 0.1,1x1x1 ] -o $sub/$ver/Segmentation.nii.gz -r 0 -p Socrates[ 1 ] -i otsu[3] --verbose 1 | grep "Elapsed time:" && \
diff -s $sub/old/Segmentation.nii.gz $sub/new/Segmentation.nii.gz && echo; done
```

Output:
```
Elapsed time: 78.2441
sub-0010 new
Elapsed time: 42.2994
Files sub-0010/old/Segmentation.nii.gz and sub-0010/new/Segmentation.nii.gz are identical

sub-0086 old
Elapsed time: 71.4629
sub-0086 new
Elapsed time: 38.4986
Files sub-0086/old/Segmentation.nii.gz and sub-0086/new/Segmentation.nii.gz are identical

sub-0101 old
Elapsed time: 73.7006
sub-0101 new
Elapsed time: 39.7073
Files sub-0101/old/Segmentation.nii.gz and sub-0101/new/Segmentation.nii.gz are identical

sub-0109 old
Elapsed time: 71.4932
sub-0109 new
Elapsed time: 38.3375
Files sub-0109/old/Segmentation.nii.gz and sub-0109/new/Segmentation.nii.gz are identical

sub-0114 old
Elapsed time: 69.5892
sub-0114 new
Elapsed time: 37.463
Files sub-0114/old/Segmentation.nii.gz and sub-0114/new/Segmentation.nii.gz are identical
```

To me the changes seems quite solid, maybe someone else should rerun my experiments or do further tests to if they get similar results.