@cookpa 2022-12-08
Can we check it remains consistent with prior probability images? That's probably the most common use case.