@ntustison 2022-12-08
Yeah, that's a good idea.  I'll try running a few subjects through the antsAtroposN4.sh pipeline.