@ntustison 2022-12-14
I've been trying to run this in the background but am having difficulty getting identical results.  I think I traced it to ``antsBrainExtraction.sh`` where there appears to be randomness even after setting ``-u 0``.  @cookpa , specifically, I'm looking [here](https://github.com/ANTsX/ANTs/blob/master/Scripts/antsBrainExtraction.sh#L469-L487) at the call to antsAI where I don't see the ``--random-seed`` option being used.  Can you confirm that I'm seeing this correctly and that this would be an issue if one calls ``antsBrainExtraction.sh ... -u 0 ... `` ?