@ntustison 2022-12-14
Yep, single-threaded.