@cookpa 2022-12-14
OK, then maybe we need additional changes, I'll take look. Meanwhile, I'll build this PR and run some tests using AAN4 directly