@ntustison 2022-12-14
That would be great.  I started out with antsAtroposN4 which is what led me to looking internally at the components.