@ntustison 2022-12-15
Okay, @cookpa , thanks for the reminder about setting the ANTS_RANDOM_SEED.  It appears that I'm now getting the identical results I was expecting.  I'll check after a couple more subjects and then we can merge this.  Thanks again @br-cpvc .   Regarding the randomization in antsBrainExtraction---do you want me to work on that or were you planning on addressing it?