@cookpa 2022-12-15
I also get reproducible results, and with two threads, the execution time is reduced by about 20%.