@cookpa 2022-12-15
> Regarding the randomization in antsBrainExtraction---do you want me to work on that or were you planning on addressing it?

I replied in the issue, I'll make the change to make antsBrainExtraction.sh use a consistent random seed throughout if "-u 0" is passed. I think that's pretty backwards compatible because most users wouldn't call "-u 0" unless they want reproducible results